+++ spec/features/teams_spec.rb
@@ -35,4 +35,53 @@ describe "Teams" do
       end
     end
   end
+
+  context 'when current user is the current team admin or the root' do
+    describe 'trying to add a existing user that is not in the current team' do
+      let!(:user)  { create :user, :with_team_and_is_admin }
+      let!(:user_to_be_added) { create :user, email: "user@example.com"}
+
+      it 'should update the team with this user' do
+        visit team_users_path(user.teams.first.slug)
+        click_link 'Add'
+
+        fill_in "user_email", with: "user@example.com"
+        click_button 'Add'
+
+        expect(current_path).to eq(team_new_enrollment_path(user.teams.first.slug))
+        expect(page).to have_text(I18n.t('teams.team_was_successfully_updated'))
+      end
+    end
+
+    describe 'trying to add an email that is not registered' do
+      let!(:user)  { create :user, :with_team_and_is_admin }
+
+      it 'shows an error message' do
+        visit team_users_path(user.teams.first.slug)
+        click_link 'Add'
+
+        fill_in "user_email", with: "user@example.com"
+        click_button 'Add'
+
+        expect(current_path).to eq(team_new_enrollment_path(user.teams.first.slug))
+        expect(page).to have_text(I18n.t('teams.user_no_found'))
+      end
+    end
+
+    describe 'trying to add a user that is not enrolled in the current team' do
+      let!(:user)  { create :user, :with_team_and_is_admin }
+      let!(:user_to_be_added) { create :user, teams: [user.teams.first], email: "user@example.com"}
+
+      it 'shows a message saying that the user is already on the team' do
+        visit team_users_path(user.teams.first.slug)
+        click_link 'Add'
+
+        fill_in "user_email", with: "user@example.com"
+        click_button 'Add'
+
+        expect(current_path).to eq(team_new_enrollment_path(user.teams.first.slug))
+        expect(page).to have_text(I18n.t('teams.user_is_already_in_this_team'))
+      end
+    end
+  end
 end