+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [Unreleased]
+### Changed
+- story type select to react component
 
 ## [1.4.2] - 2017-05-24
 ### Removed
@@ -17,7 +19,7 @@
 - Basic style for Labels
 - Slack Support
 - Release date to Story
-- Project now uses yarn 
+- Project now uses yarn
 
 ### Changed
 - Story id, location and history input-form-group into a react component