+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Changed
 - story type select to react component
+- Story state select into react component
 
 ## [1.4.2] - 2017-05-24
 ### Removed