+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Story state select into react component
 - Story requested_by select to react component
 - Move select components to renderSelects function
+- Alter route button Manage Members Team and associate user in team
 
 ## [1.4.2] - 2017-05-24
 ### Removed