+++ CHANGELOG.md
@@ -1,12 +1,14 @@
 # Changelog
 
 ## [Unreleased]
+
+## [1.5.0] 2017-06-23
 ### Added
 - Story History is now able to listen for changes on the story
 - Codeclimate file with engines and removed files from analysis
 
 ### Changed
-- story type select to react component
+- Story type select to react component
 - Story state select into react component
 - Story requested_by select to react component
 - Move select components to renderSelects function
@@ -109,7 +111,7 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com)
 and this project adheres to [Semantic Versioning](http://semver.org).
 
-[Unreleased]: https://github.com/Codeminer42/cm42-central/compare/v1.4.2...HEAD
+[Unreleased]: https://github.com/Codeminer42/cm42-central/compare/v1.5.0...HEAD
 [1.0.0]: https://github.com/Codeminer42/cm42-central/tree/v1.0.0
 [1.1.0]: https://github.com/Codeminer42/cm42-central/tree/v1.1.0
 [1.1.1]: https://github.com/Codeminer42/cm42-central/tree/v1.1.1
@@ -120,3 +122,4 @@ and this project adheres to [Semantic Versioning](http://semver.org).
 [1.4.0]: https://github.com/Codeminer42/cm42-central/tree/v1.4.0
 [1.4.1]: https://github.com/Codeminer42/cm42-central/tree/v1.4.1
 [1.4.2]: https://github.com/Codeminer42/cm42-central/tree/v1.4.2
+[1.5.0]: https://github.com/Codeminer42/cm42-central/tree/v1.5.0